requirements.txt
# Core LangChain and LangGraph
langchain>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.0.5
langchain-google-genai>=0.0.5

# LangSmith for tracing and evaluation
langsmith>=0.0.77

# Data processing and ML
torch>=2.0.0
pytorch-lightning>=2.0.0
torchgeo>=0.5.0
albumentations>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Geospatial libraries
rasterio>=1.3.0
geopandas>=0.14.0
shapely>=2.0.0
pystac-client>=0.7.0
odc-stac>=0.3.0

# Document processing
PyMuPDF>=1.23.0
pypdf>=3.17.0
python-docx>=1.0.0

# Dataset access
datasets>=2.14.0
kaggle>=1.5.16

# Environment and security
docker>=6.1.0
bandit>=1.7.5
pip-audit>=2.6.0

# Utilities
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
